+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 # 1.1.99
 
 - Crunchyroll: fix bug cancel progress if intro at second 0.
+- Amazon: fixed speed slider for A/B test UI.
 
 ## 1.1.98